Move agent-session-event.ts to lib
